+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Add `required_pkgs()` methods to `int_conformal_cv()`, `int_conformal_full()`, `int_conformal_quantile()`,  and `int_conformal_split()`. (#190)
 
+* Add butcher methods to `int_conformal_cv()`, `int_conformal_full()`, `int_conformal_quantile()`,  and `int_conformal_split()`. (#194)
+
 # probably 1.1.1
 
 * Updated unit tests for new ggplot2 release (#180).